rust/hww: fix module-inception
api::api is bad (clippy), so we move it to mod.rs.
